Skip corrupt javadoc jar files
Fixes https://github.com/jenkins-infra/helpdesk/issues/3290

https://github.com/jenkins-infra/helpdesk/issues/3290 reports that javadoc
builds have been failing since the release of the zdevops plugin 1.0.0.
Core javadoc is still documenting Jenkins 2.380 even though we have
released 2.381 and 2.382.

When the javadoc zip file is empty, we can safely ignore it and continue
to generate the rest of the javadoc site.  No need to crash the javadoc
site generation just because a damaged or flawed javadoc jar file has
been published.

It is unfortunate that @IBA-mainframe-dev has not responded to the issue,
but the javadoc site creator script is better if it handles corrupt files
without crashing.